+++ corehq/apps/commtrack/parsing.py
@@ -57,7 +57,7 @@ from casexml.apps.stock import const as stockconst
 from casexml.apps.stock.const import COMMTRACK_REPORT_XMLNS
 from corehq.apps.commtrack import const
 from corehq.apps.commtrack.exceptions import InvalidDate
-from corehq.form_processor.parsers.ledgers import StockReportHelper, StockTransactionHelper
+from corehq.form_processor.parsers.ledgers.helpers import StockReportHelper, StockTransactionHelper
 from couchforms.models import XFormInstance
 from corehq.form_processor.utils import adjust_datetimes
 from xml2json.lib import convert_xml_to_json
