+++ corehq/apps/commtrack/tests/test_stock_report.py
@@ -11,7 +11,7 @@ from casexml.apps.stock.const import REPORT_TYPE_BALANCE
 from casexml.apps.stock.models import StockReport, StockTransaction
 from corehq.apps.commtrack.processing import create_models_for_stock_report
 from corehq.apps.domain.shortcuts import create_domain
-from corehq.form_processor.parsers.ledgers import StockReportHelper, StockTransactionHelper as STrans
+from corehq.form_processor.parsers.ledgers.helpers import StockReportHelper, StockTransactionHelper
 from couchforms.models import XFormInstance
